documentation

+++ include/angle.hpp
@@ -7,27 +7,29 @@
 
 namespace ts
 {
-    //
+    /// \brief class representing a geometric angle
     struct Angle
     {
         public:
-            //
+            /// \brief convert to degrees
+            /// \returns float, in [0, 360]
             float as_degrees() const;
 
-            //
+            /// \brief convert to radians
+            /// \returns float, in [pi, 2pi]
             float as_radians() const;
 
         private:
             Angle(float dg);
             float _degrees;
 
-            friend Angle degrees(int);
+            friend Angle degrees(float);
             friend Angle radians(float);
     };
 
-    //
-    Angle degrees(int);
+    /// \brief construct angle in degree
+    Angle degrees(float);
 
-    //
+    /// \brief construct angle in radian
     Angle radians(float);
 }
